+++ src/cutadapt/_align.pyx
@@ -86,7 +86,7 @@ class DPMatrix:
 	"""
 	Representation of the dynamic-programming matrix.
 
-	This used only when debugging is enabled in the Aligner class since the
+	This is used only when debugging is enabled in the Aligner class since the
 	matrix is normally not stored in full.
 
 	Entries in the matrix may be None, in which case that value was not
@@ -226,7 +226,7 @@ cdef class Aligner:
 		"""
 		def __set__(self, value):
 			if value < 1:
-				raise ValueError('Insertion/deletion cost must be at leat 1')
+				raise ValueError('Insertion/deletion cost must be at least 1')
 			self._insertion_cost = value
 			self._deletion_cost = value